feat(session)!: establish wp2.1 domain contracts
Add immutable generic command outcomes and restrict session rulesets to
Classic and Crapless.

Define the Bubble Craps vig policy, use float monetary values, reconcile
deferred GameSession commands, and publish the WP2.1 checkpoint version.

BREAKING CHANGE: SessionConfiguration removes table_settings and
casino_profile, requires a Ruleset value, and uses float bankroll values.
Roll and shooter records now use aggregate float monetary contracts, and
GameSession command shells return CommandResult.

+++ src/bubblecraps/session/state.py
@@ -38,7 +38,7 @@ class GameState:
 
     phase: GamePhase
     actions: AvailableActions
-    bankroll: int
+    bankroll: float
     point: int | None
     puck_on: bool
     bets: list[Bet]
